Extract editor manifest parsing service
Moved all `*.editor.toml` parsing and validation logic out of `PackageManifestLoader` into a new `EditorManifestLoader`, keeping package loading focused on package-level metadata and contribution wiring. Added a shared `TomlTableReader` utility for typed TOML accessors and updated package loading to use it, reducing duplication while preserving existing manifest validation behavior.
